+++ actstream/gfk.py
@@ -1,4 +1,3 @@
-import django
 from django.conf import settings
 from django.db.models import Manager
 from django.db.models.query import QuerySet, EmptyQuerySet
@@ -7,10 +6,6 @@ from django.utils.encoding import smart_unicode
 from django.contrib.contenttypes.models import ContentType
 from django.contrib.contenttypes.generic import GenericForeignKey
 
-USE_PREFETCH = getattr(settings, 'USE_PREFETCH',
-                       django.VERSION[0] == 1 and django.VERSION[1] >= 4)
-FETCH_RELATIONS = getattr(settings, 'FETCH_RELATIONS', True)
-GFK_FETCH_DEPTH = getattr(settings, 'GFK_FETCH_DEPTH', 0)
 
 class GFKManager(Manager):
     """
@@ -37,9 +32,11 @@ class GFKQuerySet(QuerySet):
     and empty querysets.
     """
     def fetch_generic_relations(self, *args):
+        from actstream import settings as actstream_settings
+
         qs = self._clone()
 
-        if not FETCH_RELATIONS:
+        if not actstream_settings.FETCH_RELATIONS:
             return qs
 
         gfk_fields = [g for g in self.model._meta.virtual_fields
@@ -47,16 +44,18 @@ class GFKQuerySet(QuerySet):
         if args:
             gfk_fields = filter(lambda g: g.name in args, gfk_fields)
 
-        if USE_PREFETCH and hasattr(self, 'prefetch_related'):
+        if actstream_settings.USE_PREFETCH and hasattr(self, 'prefetch_related'):
             return qs.prefetch_related(*[g.name for g in gfk_fields])
 
         ct_map, data_map = {}, {}
 
         for item in qs:
             for gfk in gfk_fields:
+                if getattr(item, gfk.fk_field) is None:
+                    continue
                 ct_id_field = self.model._meta.get_field(gfk.ct_field).column
-                if getattr(item, ct_id_field) is None: continue
-                if getattr(item, gfk.fk_field) is None: continue
+                if getattr(item, ct_id_field) is None:
+                    continue
                 ct_map.setdefault(getattr(item, ct_id_field), {}
                     )[smart_unicode(getattr(item, gfk.fk_field))] = (gfk.name,
                         item.pk)
@@ -68,7 +67,7 @@ class GFKQuerySet(QuerySet):
                 ct = ctypes[ct_id]
                 model_class = ct.model_class()
                 objects = model_class._default_manager.select_related(
-                    depth=GFK_FETCH_DEPTH)
+                    depth=actstream_settings.GFK_FETCH_DEPTH)
                 for o in objects.filter(pk__in=items_.keys()):
                     (gfk_name, item_id) = items_[smart_unicode(o.pk)]
                     data_map[(ct_id, smart_unicode(o.pk))] = o
@@ -78,12 +77,11 @@ class GFKQuerySet(QuerySet):
                 if getattr(item, gfk.fk_field) != None:
                     ct_id_field = self.model._meta.get_field(gfk.ct_field)\
                         .column
-                    ctype = getattr(item, ct_id_field)
-                    gfk_pk = smart_unicode(getattr(item, gfk.fk_field))
-                    key = smart_unicode(getattr(item, gfk.fk_field))
-                    if key in data_map:
-                        setattr(item, gfk.name, data_map[key])
-                    # If the value isn't found, we leave it as is
+                    setattr(item, gfk.name,
+                        data_map[(
+                            getattr(item, ct_id_field),
+                            smart_unicode(getattr(item, gfk.fk_field))
+                        )])
 
         return qs