+++ docs/source/conf.py
@@ -10,16 +10,16 @@
 #
 # All configuration values have a default; values that are commented out
 # serve to show the default.
-
-import sys
 import os
 from datetime import datetime
 
-sys.path.insert(0, os.path.abspath(os.path.join('..', '..', 'actstream', 'runtests')))
+os.environ['DJANGO_SETTINGS_MODULE'] = 'actstream.runtests.settings'
 
-import settings
-from django.core.management import setup_environ
-setup_environ(settings)
+import django
+try:
+    django.setup()
+except AttributeError:
+    pass
 
 import actstream
 
@@ -35,7 +35,7 @@ import actstream
 
 # Add any Sphinx extension module names here, as strings. They can be extensions
 # coming with Sphinx (named 'sphinx.ext.*') or your custom ones.
-extensions = ['sphinx.ext.autodoc']
+extensions = ['sphinx.ext.autodoc', 'sphinx.ext.viewcode']
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
@@ -130,7 +130,7 @@ html_theme = 'default'
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+#html_static_path = ['_static']
 
 # If not '', a 'Last updated on:' timestamp is inserted at every page bottom,
 # using the given strftime format.